# Conflicts:
#	Minigames/src/main/java/au/com/mineauz/minigames/backend/mysql/MySQLBackend.java
#	Minigames/src/main/java/au/com/mineauz/minigames/backend/sqlite/SQLiteBackend.java
#	Minigames/src/main/java/au/com/mineauz/minigames/commands/ScoreboardCommand.java
#	Minigames/src/main/java/au/com/mineauz/minigames/commands/set/SetBlockWhitelistCommand.java
#	Minigames/src/main/java/au/com/mineauz/minigames/commands/set/SetFloorDegeneratorCommand.java
#	Minigames/src/main/java/au/com/mineauz/minigames/commands/set/SetTimerCommand.java
#	Minigames/src/main/java/au/com/mineauz/minigames/managers/language/langkeys/MgMenuLangKey.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemBoolean.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemComponent.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemCustom.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemDisplayEffects.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemDisplayLoadout.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemDisplayWhitelist.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemEnum.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemInteger.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemList.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemLong.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemModifyStatSetting.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemNewLine.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemRewardAdd.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemRewardGroup.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemSaveMinigame.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemString.java
#	Minigames/src/main/java/au/com/mineauz/minigames/menu/MenuItemWhitelistBlock.java
#	Minigames/src/main/java/au/com/mineauz/minigames/minigame/Minigame.java
#	Minigames/src/main/java/au/com/mineauz/minigames/minigame/ScoreboardDisplay.java
#	Minigames/src/main/java/au/com/mineauz/minigames/minigame/modules/TeamsModule.java
#	Minigames/src/main/java/au/com/mineauz/minigames/presets/PresetLoader.java
#	Minigames/src/main/java/au/com/mineauz/minigames/signs/RewardSign.java
#	Minigames/src/main/java/au/com/mineauz/minigames/signs/ScoreSign.java
#	Minigames/src/main/java/au/com/mineauz/minigames/signs/TeamSign.java
#	Minigames/src/main/java/au/com/mineauz/minigames/stats/StatSettings.java
#	Minigames/src/main/java/au/com/mineauz/minigames/tool/MinigameTool.java
#	Minigames/src/main/resources/messages.properties
#	Regions/src/main/java/au/com/mineauz/minigamesregions/actions/SpawnEntityAction.java
#	Regions/src/main/java/au/com/mineauz/minigamesregions/conditions/MatchBlockCondition.java
#	Regions/src/main/java/au/com/mineauz/minigamesregions/conditions/MatchTeamCondition.java
#	Regions/src/main/java/au/com/mineauz/minigamesregions/conditions/PlayerHasItemCondition.java
#	Regions/src/main/java/au/com/mineauz/minigamesregions/menuitems/MenuItemNodeExecutor.java
#	Regions/src/main/java/au/com/mineauz/minigamesregions/menuitems/MenuItemRegionExecutor.java

+++ Minigames/src/main/java/au/com/mineauz/minigames/recorder/RollbackScheduler.java
@@ -23,7 +23,7 @@ public class RollbackScheduler implements Runnable {
         iterator = blocks.iterator();
         this.minigame = minigame;
         this.modifier = modifier;
-        int delay = minigame.getRegenDelay() * 20 + 1;
+        long delay = minigame.getRegenDelay() * 20 + 1;
         task = Bukkit.getScheduler().runTaskTimer(Minigames.getPlugin(), this, delay, 1);
     }
